// apps/backend/src/prisma/migrations/snapshots/abd6ea91d9db085081451075f74b7ccc4f39b6a7fa201cf7611a90a2cb9cdf74/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Department: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly department:
        'computer science' | 'software engineering' | 'data science' | 'cybersecurity';
      readonly max_level: CodecTypes['pg/int4@1']['output'];
    };
    readonly Document: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly documentType: CodecTypes['pg/text@1']['output'];
      readonly fileName: CodecTypes['pg/text@1']['output'];
      readonly filePath: CodecTypes['pg/text@1']['output'];
      readonly status: 'pending' | 'approved' | 'rejected';
      readonly uploadDate: CodecTypes['pg/date-temporal@1']['output'];
      readonly reviewDate: CodecTypes['pg/date-temporal@1']['output'] | null;
      readonly comment: CodecTypes['pg/text@1']['output'] | null;
      readonly lockedBy: CodecTypes['pg/int4@1']['output'] | null;
      readonly lockedAt: CodecTypes['pg/date-temporal@1']['output'] | null;
      readonly studentId: CodecTypes['pg/int4@1']['output'];
      readonly staffId: CodecTypes['pg/int4@1']['output'];
      readonly sessionId: CodecTypes['pg/int4@1']['output'];
    };
    readonly Notification: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly message: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/date-temporal@1']['output'];
      readonly isRead: CodecTypes['pg/bool@1']['output'];
      readonly studentId: CodecTypes['pg/int4@1']['output'];
      readonly staffId: CodecTypes['pg/int4@1']['output'];
    };
    readonly Registration: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly status: 'not started' | 'ongoing' | 'completed' | 'failed';
      readonly dateRegistered: CodecTypes['pg/date-temporal@1']['output'] | null;
      readonly studentId: CodecTypes['pg/int4@1']['output'];
      readonly sessionId: CodecTypes['pg/int4@1']['output'];
    };
    readonly requiredDocuments: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
    };
    readonly requiredDocumentsMAPstudentCategories: {
      readonly stuCategoryId: CodecTypes['pg/int4@1']['output'];
      readonly reqDocsId: CodecTypes['pg/int4@1']['output'];
    };
    readonly Session: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly sessionId: CodecTypes['pg/text@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly startDate: CodecTypes['pg/date-temporal@1']['output'];
      readonly endDate: CodecTypes['pg/date-temporal@1']['output'] | null;
    };
    readonly Staff: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly staffId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly Student: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly matric_no: CodecTypes['pg/text@1']['output'];
      readonly department:
        'computer science' | 'software engineering' | 'data science' | 'cybersecurity';
      readonly level: CodecTypes['pg/int4@1']['output'];
      readonly mode_of_entry: 'utme' | 'transfer' | 'direct entry';
      readonly graduated: CodecTypes['pg/bool@1']['output'];
      readonly categoryId: CodecTypes['pg/int4@1']['output'] | null;
      readonly academicSession: CodecTypes['pg/text@1']['output'] | null;
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly departmentId: CodecTypes['pg/int4@1']['output'];
    };
    readonly studentCategories: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly category: 'fresher' | 'returning' | 'finalist';
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly firstName: CodecTypes['pg/text@1']['output'];
      readonly lastName: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/date-temporal@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly mustResetPassword: CodecTypes['pg/bool@1']['output'];
      readonly passwordRestAt: CodecTypes['pg/date-temporal@1']['output'] | null;
      readonly role: 'student' | 'staff' | 'admin';
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Department: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly department:
        'computer science' | 'software engineering' | 'data science' | 'cybersecurity';
      readonly max_level: CodecTypes['pg/int4@1']['input'];
    };
    readonly Document: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly documentType: CodecTypes['pg/text@1']['input'];
      readonly fileName: CodecTypes['pg/text@1']['input'];
      readonly filePath: CodecTypes['pg/text@1']['input'];
      readonly status: 'pending' | 'approved' | 'rejected';
      readonly uploadDate: CodecTypes['pg/date-temporal@1']['input'];
      readonly reviewDate: CodecTypes['pg/date-temporal@1']['input'] | null;
      readonly comment: CodecTypes['pg/text@1']['input'] | null;
      readonly lockedBy: CodecTypes['pg/int4@1']['input'] | null;
      readonly lockedAt: CodecTypes['pg/date-temporal@1']['input'] | null;
      readonly studentId: CodecTypes['pg/int4@1']['input'];
      readonly staffId: CodecTypes['pg/int4@1']['input'];
      readonly sessionId: CodecTypes['pg/int4@1']['input'];
    };
    readonly Notification: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly message: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/date-temporal@1']['input'];
      readonly isRead: CodecTypes['pg/bool@1']['input'];
      readonly studentId: CodecTypes['pg/int4@1']['input'];
      readonly staffId: CodecTypes['pg/int4@1']['input'];
    };
    readonly Registration: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly status: 'not started' | 'ongoing' | 'completed' | 'failed';
      readonly dateRegistered: CodecTypes['pg/date-temporal@1']['input'] | null;
      readonly studentId: CodecTypes['pg/int4@1']['input'];
      readonly sessionId: CodecTypes['pg/int4@1']['input'];
    };
    readonly requiredDocuments: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly isActive: CodecTypes['pg/bool@1']['input'];
    };
    readonly requiredDocumentsMAPstudentCategories: {
      readonly stuCategoryId: CodecTypes['pg/int4@1']['input'];
      readonly reqDocsId: CodecTypes['pg/int4@1']['input'];
    };
    readonly Session: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly sessionId: CodecTypes['pg/text@1']['input'];
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly startDate: CodecTypes['pg/date-temporal@1']['input'];
      readonly endDate: CodecTypes['pg/date-temporal@1']['input'] | null;
    };
    readonly Staff: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly staffId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly Student: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly matric_no: CodecTypes['pg/text@1']['input'];
      readonly department:
        'computer science' | 'software engineering' | 'data science' | 'cybersecurity';
      readonly level: CodecTypes['pg/int4@1']['input'];
      readonly mode_of_entry: 'utme' | 'transfer' | 'direct entry';
      readonly graduated: CodecTypes['pg/bool@1']['input'];
      readonly categoryId: CodecTypes['pg/int4@1']['input'] | null;
      readonly academicSession: CodecTypes['pg/text@1']['input'] | null;
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly departmentId: CodecTypes['pg/int4@1']['input'];
    };
    readonly studentCategories: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly category: 'fresher' | 'returning' | 'finalist';
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly firstName: CodecTypes['pg/text@1']['input'];
      readonly lastName: CodecTypes['pg/text@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/date-temporal@1']['input'];
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly mustResetPassword: CodecTypes['pg/bool@1']['input'];
      readonly passwordRestAt: CodecTypes['pg/date-temporal@1']['input'] | null;
      readonly role: 'student' | 'staff' | 'admin';
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;